extracting JSON Schema definitions and code into mds-json-schema
