+++ packages/mds-compliance/package.json
@@ -23,7 +23,6 @@
   "author": "City of Los Angeles.",
   "license": "Apache-2.0",
   "dependencies": {
-    "@hapi/joi": "15.1.1",
     "@mds-core/mds-api-server": "0.1.24",
     "@mds-core/mds-cache": "0.1.24",
     "@mds-core/mds-db": "0.1.24",
@@ -31,6 +30,7 @@
     "@mds-core/mds-stream": "0.1.24",
     "@mds-core/mds-types": "0.1.21",
     "@mds-core/mds-utils": "0.1.24",
+    "@mds-core/mds-json-schema": "0.1.1",
     "express": "4.17.1",
     "fs": "0.0.1-security",
     "moment-timezone": "0.5.27",
