+++ packages/mds-json-schema/index.ts
@@ -1,2 +1,2 @@
+export * from './validators'
 export * from './exceptions'
-export * from './utils'
