+++ packages/mds-json-schema/package.json
@@ -1,0 +1,29 @@
+{
+  "name": "@mds-core/mds-json-schema",
+  "version": "0.1.1",
+  "description": "Mobility Data Specification JSON schema definition and validation",
+  "keywords": [
+    "mds",
+    "json",
+    "schema"
+  ],
+  "author": "City of Los Angeles",
+  "license": "Apache-2.0",
+  "dependencies": {
+    "@hapi/joi": "15.1.1",
+    "joi-to-json-schema": "5.1.0",
+    "@mds-core/mds-types": "0.1.21",
+    "@mds-core/mds-utils": "0.1.24"
+  },
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "files": [
+    "dist/"
+  ],
+  "scripts": {
+    "build": "tsc --build tsconfig.build.json",
+    "test": "yarn test:eslint && yarn test:unit",
+    "test:eslint": "eslint --fix --ignore-path ../../.gitignore '**/*.ts'",
+    "test:unit": "DOTENV_CONFIG_PATH=../../.env nyc --check-coverage --exclude tests --extension .ts --lines 35 --reporter=text --reporter=html ts-mocha --project ../../tsconfig.json --require tsconfig-paths/register --require source-map-support/register --require dotenv/config --recursive --timeout 5000 tests/**/*.ts"
+  }
+}
