+++ packages/mds-json-schema/tsconfig.build.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.build.json",
+  "compilerOptions": {
+    "outDir": "dist"
+  },
+  "references": [{ "path": "../../packages/mds-types/tsconfig.build.json" }]
+}
