+++ packages/mds-policy/api.ts
@@ -18,7 +18,8 @@ import express from 'express'
 // import { isProviderId, providerName } from '@mds-core/mds-providers'
 import { Policy, UUID } from '@mds-core/mds-types'
 import db from '@mds-core/mds-db'
-import { now, pathsFor, NotFoundError, isUUID, policySchemaJson } from '@mds-core/mds-utils'
+import { now, pathsFor, NotFoundError, isUUID } from '@mds-core/mds-utils'
+import { policySchemaJson } from '@mds-core/mds-json-schema'
 import log from '@mds-core/mds-logger'
 import { PolicyApiRequest, PolicyApiResponse } from './types'
