+++ packages/mds-utils/exceptions.ts
@@ -1,11 +1,11 @@
-class BaseError extends Error {
+export class BaseError extends Error {
   public constructor(public name: string, public reason?: string, public info?: unknown) {
     super(reason)
     Error.captureStackTrace(this, BaseError)
   }
 }
 
-const reason = (error?: Error | string) => (error instanceof Error ? error.message : error)
+export const reason = (error?: Error | string) => (error instanceof Error ? error.message : error)
 
 /* istanbul ignore next */
 export class ServerError extends BaseError {
@@ -39,13 +39,6 @@ export class AuthorizationError extends BaseError {
 export class RuntimeError extends BaseError {
   public constructor(error?: Error | string, public info?: unknown) {
     super('RuntimeError', reason(error))
-  }
-}
-
-/* istanbul ignore next */
-export class ValidationError extends BaseError {
-  public constructor(error?: Error | string, public info?: unknown) {
-    super('ValidationError', reason(error))
   }
 }
